pyproject.toml
[project]
name = "sentientagent_v2"
version = "0.1.0"
description = "A lightweight skills-first agent built with Google ADK"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.25.0",
    "greenlet>=3.1.1",
]

[project.optional-dependencies]
feishu = [
    "lark-oapi>=1.4.15",
    "python-socks>=2.5.3",
]
openai = [
    "litellm>=1.75.5,<1.80.17",
    "openai>=1.100.2",
]

[project.scripts]
sentientagent_v2 = "sentientagent_v2.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["sentientagent_v2"]

[tool.hatch.build]
include = [
    "sentientagent_v2/**/*.py",
    "sentientagent_v2/skills/**",
]
